Split the economy into moonshine and money
Buying a still with moonshine made no sense — you sell the product and
pay with roubles. The loop is now distil -> sell -> expand, and three
mechanics fall out of that split which one currency could never support.

The tank has a capacity. Fill it and the stills stop, which is honest
and gives a reason to come back: while you are away the tank fills, you
return and sell it.

The price moves. A slow deterministic wave swings it +/-35% around 100
roubles a litre, so selling becomes a decision rather than a chore. It
is a pure function of time so that offline and online never disagree.

Quality upgrades now raise the price per litre instead of the volume.
Upgrades used to be interchangeable "x2 output"; they now compete on
four separate axes — output, price, tank capacity and hand yield — and
the list grew from 13 to 21.

Save format v3 migrates v2 by treating the stored moonshine as already
sold at base price. 25 new tests cover the tank ceiling, selling,
quality pricing and the market wave staying bounded and deterministic.

+++ lib/ui/widgets/shop.dart
@@ -134,7 +134,7 @@ class StillRow extends StatelessWidget {
             ),
           ),
           const SizedBox(width: GS.s3),
-          BuyButton(label: Fmt.short(cost), affordable: affordable, onTap: onBuy),
+          BuyButton(label: Fmt.money(cost), affordable: affordable, onTap: onBuy),
         ],
       ),
     );
@@ -284,7 +284,7 @@ class UpgradeRow extends StatelessWidget {
             if (purchased)
               Text('куплено', style: GType.label().copyWith(color: GColors.green))
             else
-              BuyButton(label: Fmt.short(cost), affordable: affordable, onTap: onBuy),
+              BuyButton(label: Fmt.money(cost), affordable: affordable, onTap: onBuy),
           ],
         ),
       ),
